Вносит правки по организации кода
